+++ src/constants/form-element.ts
@@ -186,11 +186,13 @@ export const OPTIONS_SOURCE_OPTIONS = [
     label: 'Custom',
     value: OptionsSource.CUSTOM,
     ariaLabel: TEST_IDS.formElementsEditor.optionsSourceOption(OptionsSource.CUSTOM),
+    icon: 'keyboard',
   },
   {
     label: 'Code',
     value: OptionsSource.CODE,
     ariaLabel: TEST_IDS.formElementsEditor.optionsSourceOption(OptionsSource.CODE),
+    icon: 'calculator-alt',
   },
 ];